Block merging unit tests now compile and are a bit cleaned up, but still fail
